nonNegativeLiteral() should be int, because many DBMS uses int32 as default for literals

+++ mug-guava/src/main/java/com/google/mu/safesql/SafeSql.java
@@ -326,7 +326,7 @@ public final class SafeSql {
    */
   public static SafeSql nonNegativeLiteral(int number) {
     checkArgument(number >= 0, "negative number disallowed: %s", number);
-    return new SafeSql(Long.toString(number));
+    return new SafeSql(Integer.toString(number));
   }
 
   /**
